Move tests and fix tox
