+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py38, py39, bioformats, czi, dv, lif, nd2, bfio, lint
+envlist = py38, py39, bioformats, czi, base-imageio, dv, lif, nd2, bfio, lint
 skip_missing_interpreters = true
 toxworkdir={env:TOX_WORK_DIR:.tox}
 
@@ -15,9 +15,11 @@ passenv =
     CI
 setenv =
     PYTHONPATH = {toxinidir}
+extras =
+    test
 deps = bioformats_jar
 commands =
-    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_bioformats_reader.py {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/readers/extra_readers/test_bioformats_reader.py {posargs}
 
 [testenv:czi]
 passenv =
@@ -26,11 +28,13 @@ passenv =
     CI
 setenv =
     PYTHONPATH = {toxinidir}
+extras =
+    test
 deps = aicspylibczi>=3.0.5
 commands =
-    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_czi_reader.py {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/readers/extra_readers/test_czi_reader.py {posargs}
 
-[testenv:default]
+[testenv:base-imageio]
 passenv =
     AWS_*
     JAVA_HOME
@@ -38,9 +42,12 @@ passenv =
 setenv =
     PYTHONPATH = {toxinidir}
 extras =
-    imageio
+    test
+    base-imageio
 commands =
-    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_default_reader.py {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/readers/extra_readers/test_default_reader.py {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/writers/extra_writers/test_timeseries_writer.py {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/writers/extra_writers/test_two_d_writer.py {posargs}
 
 [testenv:dv]
 passenv =
@@ -50,9 +57,10 @@ passenv =
 setenv =
     PYTHONPATH = {toxinidir}
 extras =
+    test
     dv
 commands =
-    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_dv_reader.py {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/readers/extra_readers/test_dv_reader.py {posargs}
 
 [testenv:lif]
 passenv =
@@ -61,9 +69,11 @@ passenv =
     CI
 setenv =
     PYTHONPATH = {toxinidir}
+extras =
+    test
 deps = readlif>=0.6.4
 commands =
-    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_lif_reader.py {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/readers/extra_readers/test_lif_reader.py {posargs}
 
 [testenv:nd2]
 passenv =
@@ -73,9 +83,10 @@ passenv =
 setenv =
     PYTHONPATH = {toxinidir}
 extras =
+    test
     nd2
 commands =
-    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_nd2_reader.py {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/readers/extra_readers/test_nd2_reader.py {posargs}
 
 [testenv:bfio]
 passenv =
@@ -85,8 +96,8 @@ passenv =
 setenv =
     PYTHONPATH = {toxinidir}
 extras =
-    bfio
     test
+    bfio
 commands =
     pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/readers/extra_readers/test_ome_tiled_tiff_reader.py {posargs}
